improvement: reorganize wasm lib
