+++ viewer/packages/pointclouds/wasm/src/shapes/mod.rs
